+++ QuickFIXn/IApplication.cs
@@ -9,8 +9,10 @@
         /// This callback provides you with a peek at the administrative messages
         /// that are being sent from your FIX engine to the counter party. This is
         /// normally not useful for an application however it is provided for any
-        /// logging you may wish to do. You may add fields in an adminstrative
+        /// logging you may wish to do. You may add fields in an administrative
         /// message before it is sent.
+        /// In some rare cases, it might be useful to throw a DoNotSend exception
+        /// in this function to suppress sending of the message by the application.
         /// </summary>
         /// <param name="message"></param>
         /// <param name="sessionID"></param>