+++ UnitTests/SessionTestSupport/MockApplication.cs
@@ -6,9 +6,14 @@ internal class MockApplication : QuickFix.IApplication
 {
     public Exception? FromAppException { get; set; }
     public Exception? FromAdminException { get; set; }
-    public QuickFix.DoNotSend? DoNotSendException { get; set; }
+    public Exception? ToAppException { get; set; }
+    public Exception? ToAdminException { get; set; }
 
-    public void ToAdmin(QuickFix.Message message, QuickFix.SessionID sessionId) { }
+    public void ToAdmin(QuickFix.Message message, QuickFix.SessionID sessionId)
+    {
+        if (ToAdminException is not null)
+            throw ToAdminException;
+    }
 
     public void FromAdmin(QuickFix.Message message, QuickFix.SessionID sessionId)
     {
@@ -18,8 +23,8 @@ internal class MockApplication : QuickFix.IApplication
 
     public void ToApp(QuickFix.Message message, QuickFix.SessionID sessionId)
     {
-        if (DoNotSendException is not null)
-            throw DoNotSendException;
+        if (ToAppException is not null)
+            throw ToAppException;
     }
 
     public void FromApp(QuickFix.Message message, QuickFix.SessionID sessionId)